libwasi: move some conversion subroutines to a separate file
